Make use of an extension method
